Added IdeasController methods getIdeas() and getIdeasOfUser()

Closes #61 and #59

See merge request kdg-ti/integration-2.2/2022-2023/team-2/youth-council-project!11

+++ src/main/java/be/kdg/youth_council_project/repository/YouthCouncilRepository.java
@@ -6,5 +6,8 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface YouthCouncilRepository extends JpaRepository<YouthCouncil, Integer> {
+public interface YouthCouncilRepository extends JpaRepository<YouthCouncil, Long> {
+
+
+
 }
